+++ tests/Doubles/Structure/property-override.php
@@ -1,0 +1,30 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Orisai\ReflectionMeta\Doubles\Structure\PropertyOverride;
+
+trait OverridePropertyParentTrait
+{
+
+	/**
+	 * Trait
+	 */
+	public string $test;
+
+}
+
+class OverridePropertyParent
+{
+
+	use OverridePropertyParentTrait;
+
+	/**
+	 * Parent
+	 */
+	public string $test;
+
+}
+
+class OverridePropertyChild extends OverridePropertyParent
+{
+
+}